+++ Device/ARM/ARMv8MBL/Source/GCC/startup_ARMv8MBL.c
@@ -3,7 +3,7 @@
  * @brief    CMSIS Core Device Startup File for
  *           ARMv8MBL Device Series
  * @version  V5.00
- * @date     25. April 2016
+ * @date     26. April 2016
  ******************************************************************************/
 /*
  * Copyright (c) 2009-2016 ARM Limited. All rights reserved.
@@ -124,7 +124,7 @@ void SPI_IRQHandler      (void) __attribute__ ((weak, alias("Default_Handler")))
  *----------------------------------------------------------------------------*/
 const pFunc __Vectors[] __attribute__ ((section(".vectors"))) = {
   /* ARMv8MBL Exceptions Handler */
-  (pFunc)&__StackTop,                       /*      Initial Stack Pointer     */
+  (pFunc)((uint32_t)&__StackTop),           /*      Initial Stack Pointer     */
   Reset_Handler,                            /*      Reset Handler             */
   NMI_Handler,                              /*      NMI Handler               */
   HardFault_Handler,                        /*      Hard Fault Handler        */
